massive updates after feedback
