+++ src/components/Footer/Footer.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import Github from "../images/github.svg";
-import Linkedin from "../images/linkedin.svg";
+import Github from "../../images/github.svg";
+import Linkedin from "../../images/linkedin.svg";
 
 const Footer = () => {
   return (
