+++ src/components/Header/header.css
@@ -1,6 +1,6 @@
 .header {
   /* background-image: url(../images/georgia-de-lotz--UsJoNxLaNo-unsplash.svg); */
-  background: url(../images/georgia-de-lotz--UsJoNxLaNo-unsplash.svg),
+  background: url(../../images/georgia-de-lotz--UsJoNxLaNo-unsplash.svg),
     lightgray 50% / cover no-repeat;
   max-width: 1440px;
   height: 576px;
@@ -11,7 +11,7 @@
 
 @media screen and (min-width: 320px) and (max-width: 767px) {
   .header {
-    background-image: url(../images/georgia-de-lotz--UsJoNxLaNo-unsplash-mobile.svg);
+    background-image: url(../../images/georgia-de-lotz--UsJoNxLaNo-unsplash-mobile.svg);
     max-width: 320px;
     height: 536px;
   }
@@ -19,7 +19,7 @@
 
 @media screen and (min-width: 768px) and (max-width: 1440px) {
   .header {
-    background-image: url(../images/georgia-de-lotz--UsJoNxLaNo-unsplash-tablet.svg);
+    background-image: url(../../images/georgia-de-lotz--UsJoNxLaNo-unsplash-tablet.svg);
     max-width: 768px;
     height: 400px;
   }
