+++ src/components/Main/Main.js
@@ -1,5 +1,5 @@
 import React from "react";
-import About from "./About";
+import About from "../About/About";
 
 const Main = () => {
   return (
