+++ src/components/ModalWithSuccess/ModalWithSuccess.js
@@ -1,5 +1,5 @@
 import React from "react";
-import closeIcon from "../images/close.svg";
+import closeIcon from "../../images/close.svg";
 
 const ModalWithSuccess = ({
   closeModal,
