+++ src/components/Search/Search.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SearchForm from "./SearchForm";
+import SearchForm from "../SearchForm/SearchForm";
 
 const Search = ({ handleSearchSubmit, setKeyword }) => {
   return (
